메뉴바 디자인

+++ src/com/sbs/example/jspCommunity/controller/UsrArticleController.java
@@ -1,4 +1,4 @@
-package com.sbs.example.jspCommunity.controller.usr;
+package com.sbs.example.jspCommunity.controller;
 
 import java.util.HashMap;
 import java.util.List;
@@ -6,16 +6,17 @@ import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.sbs.example.jspCommunity.container.Container;
 import com.sbs.example.jspCommunity.dto.Article;
 import com.sbs.example.jspCommunity.dto.Board;
 import com.sbs.example.jspCommunity.service.ArticleService;
 
-public class ArticleController {
+public class UsrArticleController {
 	private ArticleService articleService;
 
-	public ArticleController() {
+	public UsrArticleController() {
 		articleService = Container.articleService;
 	}
 
@@ -49,19 +50,36 @@ public class ArticleController {
 	}
 
 	public String showWrite(HttpServletRequest req, HttpServletResponse resp) {
+
+		HttpSession session = req.getSession();
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			req.setAttribute("alertMsg", "로그인 후 이용주세요");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
 		int boardId = Integer.parseInt(req.getParameter("boardId"));
 
 		Board board = articleService.getBoardById(boardId);
 		req.setAttribute("board", board);
 		req.getParameter("id");
 		req.getParameter("name");
-		
-		
+
 		return "usr/article/write";
 	}
 
 	public String doWrite(HttpServletRequest req, HttpServletResponse resp) {
-		int memberId = Integer.parseInt(req.getParameter("memberId"));
+
+		HttpSession session = req.getSession();
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			req.setAttribute("alertMsg", "로그인 후 이용주세요");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
+		int memberId = (int) session.getAttribute("loginedMemberId");
 		int boardId = Integer.parseInt(req.getParameter("boardId"));
 		String title = req.getParameter("title");
 		String body = req.getParameter("body");
@@ -79,12 +97,18 @@ public class ArticleController {
 		return "common/redirect";
 	}
 
-
-
-
 	public String doDelete(HttpServletRequest req, HttpServletResponse resp) {
+
+		HttpSession session = req.getSession();
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			req.setAttribute("alertMsg", "로그인 후 이용주세요");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
 		int id = Integer.parseInt(req.getParameter("id"));
-		
+
 		Article article = articleService.getForPrintArticleById(id);
 
 		if (article == null) {
@@ -92,26 +116,52 @@ public class ArticleController {
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
-		
+
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+
+		if (article.getMemberId() != loginedMemberId) {
+			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
 		articleService.delete(id);
-		
+
 		int boardId = article.getBoardId();
-		
+
 		req.setAttribute("alertMsg", id + "번 게시물이 삭제되었습니다.");
 		req.setAttribute("replaceUrl", String.format("list?boardId=%d", boardId));
 		return "common/redirect";
 	}
+
 	public String showModify(HttpServletRequest req, HttpServletResponse resp) {
+
+		HttpSession session = req.getSession();
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			req.setAttribute("alertMsg", "로그인 후 이용주세요");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
 		int id = Integer.parseInt(req.getParameter("id"));
 
 		Article article = articleService.getForPrintArticleById(id);
-		
+
 		if (article == null) {
 			req.setAttribute("alertMsg", id + "번 게시물은 존재하지 않습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
-		
+
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+
+		if (article.getMemberId() != loginedMemberId) {
+			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
 		Board board = articleService.getBoardById(article.getBoardId());
 
 		req.setAttribute("article", article);
@@ -121,24 +171,33 @@ public class ArticleController {
 	}
 
 	public String doModify(HttpServletRequest req, HttpServletResponse resp) {
+
+		HttpSession session = req.getSession();
+
+		if (session.getAttribute("loginedMemberId") == null) {
+			req.setAttribute("alertMsg", "로그인 후 이용주세요");
+			req.setAttribute("historyBack", true);
+			return "common/redirect";
+		}
+
 		int id = Integer.parseInt(req.getParameter("id"));
-		
+
 		Article article = articleService.getForPrintArticleById(id);
-		
+
 		if (article == null) {
 			req.setAttribute("alertMsg", id + "번 게시물은 존재하지 않습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
-		
-		int memberId = Integer.parseInt(req.getParameter("memberId"));
-		
-		if ( article.getMemberId() != memberId ) {
+
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+
+		if (article.getMemberId() != loginedMemberId) {
 			req.setAttribute("alertMsg", id + "번 게시물에 대한 권한이 없습니다.");
 			req.setAttribute("historyBack", true);
 			return "common/redirect";
 		}
-		
+
 		String title = req.getParameter("title");
 		String body = req.getParameter("body");
